[Need notice]
New method NexT.utils.activateSidebarPanel
292b8a258edc5176448c5a3e8ebc53448c085eaa

原修改在 source/js/next-boot.js 中进行
但在 292b8a2 中，对应部分被移动到 source/js/utils.js 中
手动合并的内容，需要注意查看。
