+++ source/js/next-boot.js
@@ -18,21 +18,7 @@ NexT.boot.registerEvents = function() {
 
   document.querySelectorAll('.sidebar-nav li').forEach((element, index) => {
     element.addEventListener('click', () => {
-      if (element.matches('.sidebar-toc-active .sidebar-nav-toc, .sidebar-overview-active .sidebar-nav-overview')) return;
-      const sidebar = document.querySelector('.sidebar-inner');
-
-      const panelContainer = sidebar.querySelector('.sidebar-panel-container');
-      const nav = panelContainer.firstElementChild.querySelector('.nav');
-      const navHeight = nav ? parseInt(nav.style.getPropertyValue('--height'), 10) : 0;
-      const panelHeights = [
-        navHeight || panelContainer.firstElementChild.scrollHeight,
-        panelContainer.lastElementChild.scrollHeight
-      ];
-      panelContainer.style.setProperty('--inactive-panel-height', `${panelHeights[1 - index]}px`);
-      panelContainer.style.setProperty('--active-panel-height', `${panelHeights[index]}px`);
-
-      const activeClassNames = ['sidebar-toc-active', 'sidebar-overview-active'];
-      sidebar.classList.replace(activeClassNames[1 - index], activeClassNames[index]);
+      NexT.utils.activateSidebarPanel(index);
     });
   });
